Activity Header data

+++ Source/FreeTrainSimulator.Models/Content/ActivityModel.cs
@@ -7,7 +7,7 @@ namespace FreeTrainSimulator.Models.Content
     [MemoryPackable(GenerateType.VersionTolerant, SerializeLayout.Sequential)]
     public sealed partial record ActivityModel : ActivityModelCore
     {
-        public int InitialSpeed { get; init; }
+        public float InitialSpeed { get; init; }
         public EnumArray<int, FuelType> FuelLevels { get; init; }
         public int HazardProbability { get; init; }
     }
